mikrotik_isp_management/conftest.py
import pytest

from mikrotik_isp_management.users.models import User
from mikrotik_isp_management.users.tests.factories import UserFactory


@pytest.fixture(autouse=True)
def media_storage(settings, tmpdir):
    settings.MEDIA_ROOT = tmpdir.strpath


@pytest.fixture
def user(db) -> User:
    return UserFactory()
